+++ src/main/java/com/samsolutions/employeesdep/model/dao/Dao.java
@@ -1,4 +1,4 @@
-package com.samsolutions.employeesdep.model.daos;
+package com.samsolutions.employeesdep.model.dao;
 
 import java.util.Optional;
 import java.util.List;
