+++ src/main/java/com/samsolutions/employeesdep/model/services/EmployeeService.java
@@ -1,6 +1,6 @@
 package com.samsolutions.employeesdep.model.services;
 
-import com.samsolutions.employeesdep.model.daos.JpaRoleDao;
+import com.samsolutions.employeesdep.model.dao.JpaRoleDao;
 import com.samsolutions.employeesdep.model.entities.Department;
 import com.samsolutions.employeesdep.model.entities.Employee;
 import com.samsolutions.employeesdep.model.entities.Role;
